Fixing the failures due r16720
Can't create money with null currency, no validation exist in 1.4.x as the work in this direction has been done in trunk. Hence it was committed unnoticed that currency was being null when Money was created.

No need for the Money(currency) in 1.4.x, the changes at trunk are different in term Money creation at MultipleLoanCreationViewHelper.

git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/branches/v1.4.x@16721 a8845c50-7012-0410-95d3-8e1449b9b1e4